게시판 #finish

+++ mysite03/src/main/java/com/poscodx/mysite/controller/BoardController.java
@@ -1,6 +1,8 @@
-package com.douzone.mysite.controller;
+package com.poscodx.mysite.controller;
 
 import java.util.Map;
+
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -10,13 +12,10 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
-
-import com.douzone.mysite.security.Auth;
-import com.douzone.mysite.security.AuthUser;
-import com.douzone.mysite.service.BoardService;
-import com.douzone.mysite.vo.BoardVo;
-import com.douzone.mysite.vo.UserVo;
-import com.douzone.web.util.WebUtil;
+import com.poscodx.mysite.service.BoardService;
+import com.poscodx.mysite.vo.BoardVo;
+import com.poscodx.mysite.vo.UserVo;
+import com.poscodx.web.util.WebUtil;
 
 @Controller
 @RequestMapping("/board")
@@ -46,34 +45,50 @@ public class BoardController {
 		return "board/view";
 	}
 	
-	@Auth
 	@RequestMapping("/delete/{no}")
 	public String delete(
-		@AuthUser UserVo authUser,
+		HttpSession session, 
 		@PathVariable("no") Long boardNo,
 		@RequestParam(value="p", required=true, defaultValue="1") Integer page,
-		@RequestParam(value="kwd", required=true, defaultValue="") String keyword) {
-		
+		@RequestParam(value="kwd", required=true, defaultValue="") String keyword) {		
+		// Access Control(접근 제어)
+		UserVo authUser = (UserVo) session.getAttribute("authUser");
+		if(authUser == null) {
+			return "redirect:/user/login";
+		}
+		///////////////////////////////////////////////////////////		
 		boardService.deleteContents(boardNo, authUser.getNo());
 		return "redirect:/board?p=" + page + "&kwd=" + WebUtil.encodeURL(keyword, "UTF-8");
 	}
 	
-	@Auth
 	@RequestMapping("/modify/{no}")	
-	public String modify(@AuthUser UserVo authUser, @PathVariable("no") Long no, Model model) {
+	public String modify(HttpSession session, @PathVariable("no") Long no, Model model) {
+		// Access Control(접근 제어)
+		UserVo authUser = (UserVo) session.getAttribute("authUser");
+		if(authUser == null) {
+			return "redirect:/user/login";
+		}
+		///////////////////////////////////////////////////////////
+		
 		BoardVo boardVo = boardService.getContents(no, authUser.getNo());
 
 		model.addAttribute("boardVo", boardVo);
 		return "board/modify";
 	}
 
-	@Auth
 	@RequestMapping(value="/modify", method=RequestMethod.POST)	
 	public String modify(
-		@AuthUser UserVo authUser,
+		HttpSession session, 
 		@ModelAttribute BoardVo boardVo,
 		@RequestParam(value="p", required=true, defaultValue="1") Integer page,
-		@RequestParam(value="kwd", required=true, defaultValue="") String keyword) {
+		@RequestParam(value="kwd", required=true, defaultValue="") String keyword) {		
+		// Access Control(접근 제어)
+		UserVo authUser = (UserVo) session.getAttribute("authUser");
+		if(authUser == null) {
+			return "redirect:/user/login";
+		}
+		///////////////////////////////////////////////////////////
+		
 		boardVo.setUserNo(authUser.getNo());
 		boardService.modifyContents(boardVo);
 		return "redirect:/board/view/" + boardVo.getNo() + 
@@ -81,31 +96,49 @@ public class BoardController {
 				"&kwd=" + WebUtil.encodeURL( keyword, "UTF-8" );
 	}
 
-	@Auth
 	@RequestMapping(value="/write", method=RequestMethod.GET)	
-	public String write() {
+	public String write(HttpSession session) {
+		// Access Control(접근 제어)
+		UserVo authUser = (UserVo) session.getAttribute("authUser");
+		if(authUser == null) {
+			return "redirect:/user/login";
+		}
+		///////////////////////////////////////////////////////////
+		
 		return "board/write";
 	}
 
-	@Auth
 	@RequestMapping(value="/write", method=RequestMethod.POST)	
 	public String write(
-		@AuthUser UserVo authUser,	
+		HttpSession session, 
 		@ModelAttribute BoardVo boardVo,
 		@RequestParam(value="p", required=true, defaultValue="1") Integer page,
 		@RequestParam(value="kwd", required=true, defaultValue="") String keyword) {
+		// Access Control(접근 제어)
+		UserVo authUser = (UserVo) session.getAttribute("authUser");
+		if(authUser == null) {
+			return "redirect:/user/login";
+		}
+		///////////////////////////////////////////////////////////
+		
 		boardVo.setUserNo(authUser.getNo());
 		boardService.addContents(boardVo);
 		
 		return	"redirect:/board?p=" + page + "&kwd=" + WebUtil.encodeURL(keyword, "UTF-8");
 	}
 
-	@Auth
 	@RequestMapping(value="/reply/{no}")	
 	public String reply(
-		@AuthUser UserVo authUser,
+		HttpSession session, 
 		@PathVariable("no") Long no,
 		Model model) {
+		// Access Control(접근 제어)
+		UserVo authUser = (UserVo) session.getAttribute("authUser");
+		if(authUser == null) {
+			return "redirect:/user/login";
+		}
+		///////////////////////////////////////////////////////////
+		
 		BoardVo boardVo = boardService.getContents(no);
 		boardVo.setOrderNo(boardVo.getOrderNo() + 1);
 		boardVo.setDepth(boardVo.getDepth() + 1);
